[NODE] Change numered list to checked list on change status function

+++ app.js
@@ -3,7 +3,8 @@ const {
   readInput,
   showTasks,
   selectTask,
-  emptyList,
+  confirmSelection,
+  deleteTask,
   pause
 } = require('./src/helpers/menus');
 const { 
@@ -30,38 +31,39 @@ const main = async() => {
         const description = await readInput('Write the task description:');
         createTask(description);
         break;
-    
       case '2': 
         const listTasksAnswer = listTasks();
-        empty = validateEmpty(listTasksAnswer, async () => {
-          await showTasks(listTasksAnswer);
-        });
+        if (empty = validateEmpty(listTasksAnswer)) break;
+        await showTasks(listTasksAnswer);
         break;
       case '3': 
         const listTasksByStateDone = listTasksByState(TaskStatus.DONE);
-        empty = validateEmpty(listTasksByStateDone, async () => {
-          await showTasks(listTasksByStateDone);
-        });
+        if (empty = validateEmpty(listTasksByStateDone)) break;
+        await showTasks(listTasksByStateDone);
         break;
       case '4': 
         const listTasksByStatePending = listTasksByState(TaskStatus.PENDING);
-        empty = validateEmpty(listTasksByStatePending, async () => {
-          await showTasks(listTasksByStatePending);
-        });
+        if (empty = validateEmpty(listTasksByStatePending)) break;
+        await showTasks(listTasksByStatePending);
         break;
       case '5':
-        const changeStatusTasks = listTasksByState(TaskStatus.PENDING);
-        empty = validateEmpty(changeStatusTasks, async () => {
-          const changeStatusTaskId = await selectTask(changeStatusTasks);
-          changeStatusById(changeStatusTaskId, TaskStatus.DONE);
-        });
+        const changeStatusTasks = listTasks();
+        if (empty = validateEmpty(changeStatusTasks)) break;
+        const changeStatusTaskIds = await selectTask(changeStatusTasks);
+        changeStatusById(changeStatusTaskIds);
         break;
       case '6': 
         const deleteTasks = listTasks();
-        empty = validateEmpty(deleteTasks, async () => {
-          const deleteTaskId = await selectTask(deleteTasks);
-          deleteTaskById(deleteTaskId);
-        });
+        if (empty = validateEmpty(deleteTasks)) break;
+        const deleteTaskId = await deleteTask(deleteTasks);
+        if (deleteTaskId != '-1') {
+          const confirm = await confirmSelection('Are you sure?');
+          if (confirm){
+            deleteTaskById(deleteTaskId);
+          }
+        } else {
+          empty = null;
+        }
         break;
     }
     await pause(empty);
@@ -69,13 +71,8 @@ const main = async() => {
   console.clear();
 }
 
-const validateEmpty = (tasks, callback) => {
-  if (Object.keys(tasks).length > 0) {
-    callback();
-  } else {
-    return true;
-  }
-  
+const validateEmpty = (tasks) => {
+  return !(Object.keys(tasks).length > 0);
 }
 
 main();
